refactor: video edit component to shared module
Moves the `VideoEditComponent` from the `video-edit` feature module to the `shared/ui` module.

This change improves code reusability and organization by placing the video edit component in a shared location, making it accessible to other modules.
