+++ packages/plugins/videos-ui/src/lib/shared/ui/video-edit/video-edit.component.scss
